Update algorithm.md
Corrected some spelling errors, and put in some "action" for the algorithm to do when the user has chosen a genre. 

I think it's good that the user gets the choice to enter whether or not it wants another suggestion, but then we need to come up with a way to write a loop? or a repetition of the process

Ill try to look into that more later after work! :)