Update algorithm.md
I tried to add a file path, but since we have the dataset stored in github, i'm not sure how to write that one. So I just guessed we can use the URL. Asked Wout about it and waiting for his answer. 

But then again, maybe I'm overcomplicating this. Since this is just a sample pseudo code, maybe the file path can just be something we can come up with. It doesn't necessary have to be a 'real' path. Maybe?