Update algorithm.md
added chosenGenre to the function filterBooksByGenre, because i saw magdalena's python version and it seems logic to include chosenGenre too. Can't believe I didn't think of that before. But then i wonder if we still need the bookList in this function? Or is it enough to filter the books by genre from the user's chosen genre? 

Sorry for the many edits sandra!